requirements.txt
SQLAlchemy==2.0.32
python-dotenv==1.0.1
pydantic==2.8.2
requests==2.32.3
# CrewAI stack
crewai==0.186.1
crewai-tools==0.0.1
# Google Gemini
google-generativeai==0.7.2
fastapi==0.112.0
uvicorn==0.30.5
python-multipart==0.0.9
PyYAML==6.0.2
passlib[bcrypt]==1.7.4
python-jose==3.3.0
